pyproject.toml
[project]
name = "take-home-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.15.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "langchain-core>=0.3.0",
    "pandas>=2.0.0",
]
